requirements.txt
asgiref==3.8.1
attrs==24.3.0
Django==5.1.4
django-filter==24.3
djangorestframework==3.15.2
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Markdown==3.7
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.22.3
sqlparse==0.5.3
uritemplate==4.1.1
